Applied refactor to Proxy tests.
